requirements.txt
aiogram
yt-dlp
ffmpeg-python
python-dotenv